When changeable features are requested, convert OR to matching OR
If at least one changeable feature is requested, then all the nodes in the
job allocation need to match the same feature set. For example, with the
following job request:

salloc -N2 -C 'a|b'

If at least one of 'a' or 'b' are a changeable feature, then all of the
nodes in the job allocation need to have either 'a' or 'b'. You cannot have
one node with 'a' and the other node with 'b'.

Since the scheduling logic always interprets OR '|' inside of parentheses as
boolean OR rather than matching OR, the feature request must be restructured
such that all OR operators are outside of parentheses and that all OR
operators are considered matching OR. For example:

'(a|b)&(c|d)'

If any of these features are changeable, then this feature expression will
be restructured internally to the following, with each '|' meaning
matching OR:

'(a&c)|(a&d)|(b&c)|(b&d)'

The original feature request string is preserved.

Bug 12286