Initial backoff support.
If the slurmctld doesn't want to talk to us at the moment, sleep for a
second and then retry the RPC again... indefinitely. This is designed
to avoid having the error out on any RPCs which would cause user-commands
to fail, while allowing the slurmctld to push back on noisy users.

Bug 14493.