slurmd - close fd when lib dir for sbcast fails to be created
fd will be open from previous call to _open_as_other() for the file that is
being transmitted. When _open_as_other() for the lib dir does not error, fd
is closed by _free_file_bcast_info_t() after it's been added to
file_bcast_list and cleanup happens. In this case though, fd never gets
added to file_bcast_list, so it needs to be closed here.

Bug 15728