Overwrite lib dir for `sbcast --send-libs` when using --force
Previously, the --force option for sbcast would not work if `--send-libs`
was also enabled. --force will replace the destination file if it exists
(O_TRUNC flag set for open()), but it will not replace the destination
library directory (mkdir() in _open_as_other() will just fail).

Using --force can be convenient for making consecutive sbcast calls in
case of failure, but this is not possible if `--send-libs` is also set.

This commit will replace the library directory if --force is set,
allowing sbcast w/ --force set to be called consecutively.

Bug 15728