Docs - Improve CoreSpec docs
Avoid confusion where it sounded like --core-spec allowed jobs to use
the CoreSpec of nodes in a way that bigger --core-spec values result
in more cores being allocated. That's not how it works.
The --core-spec options only allows to override the CoreSpecCount and
CPUSpecList of nodes, and does an implicit --exclusive.
So, the lower value of --core-spec the more cores the job will use.

Bug 15785