Don't rebuild the feature string when validating
There is no need to rebuild the feature string.
In addition, when rebuilding the feature string, the brackets and
parentheses could get placed incorrectly, resulting in a confusing log
message. For example, the following feature expression is invalid
because it lacks counts inside of brackets, but the string is rebuilt
incorrectly:

'[(a|b)&c]'

slurmctld log:
Job specs has invalid feature list: (a|[b)&c]

Bug 15715