Always requeue job if prolog failed and prolog flag ForceRequeueOnFail set
If ForceRequeueOnFail is set in PrologFlags in slurm.conf and a prolog
launch failure was detected by slurmctld, always requeue the job even if
the job was submitted with requeueing disabled.  This enables an admin to
set ForceRequeueOnFail in slurm.conf and always cause a job to be requeued
when there is a prolog failure.

Bug 11499