Fix several races with job_container/tmpfs, X11, Cray and pam_slurm_adopt
Since tmpfs adds every job step to a private namespace for /tmp, the
.Xauthority file must be created after joining the namespace. This was done
after commit 2ba2a11dc879, which introduced a race with cncu plugin.

In cncu the container_g_join moved the sleep process into a new cray job
which caused the parent cray job to be deleted. It caused issues with
pam_slurm_adopt in cray where pids could not be adopted, plus it left
stray sleep and extern processes behind.

Moreover 2ba2a11dc879 introduced a potential race if xauth command was
slower than srun and it started an x11 app causing the job to crash This
commit also hid an incorrect x11_signal_handler_thread cleanup which
caused not to wait for the thread to end with possibly getting errors while
trying to cleanup cgroups or other facilities.

Bug 15074