Fix leaving job in resume_job_list
Jobs were unnecessarily being left on the resume_job_list when they
could be removed after resuming all nodes for the job. Wasn't causing a
problem because it would be cleared the next time because either the job
wasn't configuring anymore or the nodes were out of the
power_node_bitmap from the previous resume.

Bug 15858