_schedule - fix handling of --prefer for job arrays
Since --prefer is now another dimension of job_queue we need to
handle it for job arrays as we do for partition and reservation.

Before going to "next task" we need to make sure that feature_list_use
is set correctly, so we don't attempt starting --prefer if job_queue_rec
we're dealing with is for use_prefer = false.

Bug 15886