Allow different cased strings in slurm_addto_char_list_with_case()
e.g.
If adding "foo, Foo, FOO" with no lower casing, the list would only have
"foo" because we do xstrcasecmp if the current string exists in the list
already. When no lower casing we want all three to be in the list.

Change to only append if the value to be appended doesn't exist in the
list.

Bug 15185