Add last_deficit to licenses_t and fold into scheduling decisions.
The "deficit" is defined as the number of licenses that are being consumed
beyond what slurmctld otherwise expects. The slurmctld expects up to it's
used count, plus the number of licenses that are not allocated to it, to
be in use at any given time - anything exceeding that is the deficit.

This value needs to persist between updates of the last_consumed field so
that we can continue to allocate additional licenses, while not immediately
exceeding the total license count defined through sacctmgr.