Add check to detect and abort invalid HetJob components.
Add a check to _validate_het_jobs() where a job with the following
characteristics is aborted:

HetJobOffset != 0 (In theory only non-leader HetJobs)
AND
JobId == HetJobId (In theory only leader HetJobs)

If a job record satisfies these two above, it is not coherent/valid.
Detect this situation and abort the job consequently, so it doesn't
remain unpurged indefinitely.

Co-authored-by: Oscar Hernández <oscar.hernandez@schedmd.com>

Bug 15610