Add logic to wait MAX_MOVE_WAIT milis instead of only retries in cgroup/v1
We were doing MAX_MOVE_WAIT times a check to see if the slurmstepd pid
was removed from the cgroup. It seems it is not enough or too fast on some
systems so change the wait to behave more like cgroup/v2, waiting a maximum
of 1000 ms by doing 10 retries every MAX_MOVE_WAIT/10 ms.

Bug 15833