Workaround for a kernel < 3.18 bug in cpuset cgroup interface
In kernels < 3.18 and without patch 41c25707d21716826e3c1f60967f5550610ec1c9
the cpuset cgroup interface is considered alive when it is dying, thus
giving a EBUSY to a userland program which tries to rmdir an apparently
empty cgroup. This were causing failures on our cgroup cleanup. Add a retry
and a sleep to try to mitigate this issue.

Bug 15833