Take into account MEM_PER_CPU flag when old message received
Since the change introduced in 21.08 to use [job|step]_mem_alloc as the
reference for the memory limitation in the credential, in case that a
message from an old version (20.11) is received, this will not have the
*_mem_alloc field, thus we need to do it the old way, by using the *_mem_limit
instead. Also checks the protocol_version to decide whether calling _get_ncpus
is needed.
This commit also changes the slurm_cred_get_mem signature in order to
receive the step and job cpu count so that the number of cpus can be
calculated.

Bug 15417