Replace one-bit bit-field with bool.
These are a holdover from when bool was not standardized.

While there is a change from pack32() to packbool() here, this is happening
going from the slurmd to the slurmstepd during startup and should not pose a
problem. The only time this change could be an issue is if the slurmd and
slurmstepd versions are mismatched between rc1 and the next 23.02 release.

Bug 16038

Signed-off-by: Tim Wickberg <tim@schedmd.com>