Calculate requested CPU count in the step with --cpus-per-gpu
If --cpus-per-gpu is requested, the gres requests first need to be
parsed. In addition, the step may inherit gres from the job, which
happens in slurmctld in _copy_job_tres_to_step().

So, update the step cpu count in the slurmctld after the step has
inherited job gres and other gres options have been parsed.

Bug 15632