data_parser/v0.0.39 - improve equal flag masking
When parsing a flag array, keep track of the equal flag masks and ensure
that equal flags and bit flags with overlapping masks do not clear existing
already set bits.

Flags are considered special and are handled by a flag array. Each flag
is can be one of two types: equal or bit. An equal means that the flag
sets all the bits in a given bit mask, meaning that if the flag is false
that all of those bits are 0. For a bit flag (usually a SLURM_BIT()),
then one or more bits are set 1 if the flag is true but the bit mask is
not used to 0 all the other bits. Slurm's api has a lot of mixes of bit
and equal flags all over the place even in the same integer. For this
bug, the setting of the 0s was not correct as the parsing code didn't
track the overlapping equal flags correctly.  The equal flags have a
priority of their declaration since in most flag integers, they overlap
at least partially. The main offender is setting the entire flag to
NO_VAL but which needs to be caught first to avoid the parser thinking
all the other bits are set to 1 (except the small endian).

Bug 16008