Handle mismatched TRES count from the slurmstepd's.
If the TRES was altered in the slurmctld before 1fbb55731d the slurmd
learning of the change did not rebuild the buffer being sent when
starting a new stepd. The result would be the stepd would start with
the wrong TRES count and upon finishing the ctld would assert on the
difference.

This patch corrects this by reindexing the mismatched values from the
stepd with those in the ctld.

Bug 15198