Do not constrain memory in task/cgroup if CR_Memory is not set
If CR_Memory is not set in slurm.conf we must take memory into account just
for filtering nodes out of the allocation, as is documented in slurm.conf.5
man page, so task/cgroup must not use memory.

This is a side effect of bug 15489 where we changed to use the configured
memory instead of the physical memory in task/cgroup. If there's no
RealMemory configured then it is assumed it is 1MB.

Bug 16056