job_container/tmpfs - cleanup container if namespace is already unmounted.
If the namespace (.ns) was already unmounted for some reason, then
job_container/tmpfs would error and not cleanup the rest of the
job directory. This is wanted for other errors that result in failing to
unmount the namespace, but in this situation where the namespace was
already unmounted, the job directory can be cleaned up.

Bug 16013