Pass user_id and group_id of job to sibling clusters
In 23.02, the client doesn't send the user_id and group_id in the
job_desc_msg_t anymore (ca4bb9504c) and is set from the auth message
when received by the origin cluster. The origin cluster needs to send
this along to the siblings now since the job_desc_msg_t is just passed
along to the sibling clusters without re-packing the job_desc.

Bug 16104