Fix memory leak in hierarchical rec hash
_arch_hash_rec_id() returns an xmalloc'ed key. uthash stores the pointer
to the key but xhash and uthash don't free the key (rightly so).
c15531c5f2 removed the reference to the key in xhash_item_t so it can't
be free'd by an xhash freefunc.

Because slurmdb_hierarchical_rec_t is used in a list and hash, it's
easier to just store the key in slurmdb_hierarchical_rec_t and free it
when it's destroyed -- rather than reverting c15531c5f2.

Bug 16118

Co-authored-by: Tim Wickberg <tim@schedmd.com>