Add EnableControllers option for cgroup/v2
Before, if CgroupAutomount + IgnoreSystemd were set, the cgroup/v2 plugin
would enable the controllers in the cgroup tree. This was not convenient
because it required IgnoreSystemd, but we might want controllers to be
manually enabled even if systemd was used to manage the cgroup tree. This
can happen for example in containerized environments if some levels of the
tree did not start with all the controllers enabled.

Add a new EnableControllers option that allows this logic to happen even if
we are using systemd and improve the logic to recursively try to enable
them through all the levels, from the root down to our leaf.

This option is best considered as a developer option.

Bug 16151