sched/backfill - fix segfault when removing a PLANNED node from system
Backfill's planned_bitmap isn't updated when a node is deleted from
node_record_table_ptr so the planned_bitmap may be pointing to a deleted
node. In 23.02, this isn't an issue because of the use of
next_node_bitmap() which will skip any nodes that are deleted at the set
bit.

Bug 16165