Fix leaving deleted nodes in planned bitmap for 23.02
Continuation of 5009c46c8b for 23.02.

In 22.05, we walked from the first bit to the last bit checking for
valid node records in the node_record_table_ptr. 23.02 was switched to
use next_node_bitmap() which skips empty node slots in
node_record_table_ptr. So it would leave the bit set if the node was
removed. Using bit_ffs_from_bit() we can clear the bit if the node
doesn't exist while still only walking set bits.

Bug 16165

Signed-off-by: Skyler Malinowski <malinowski@schedmd.com>