Add explicit_scancel to ScronParameters.
This option will prevent a bare scancel from cancelling a job that was
submitted with scrontab. It is somewhat easy for a user to mistakenly
cancel the job when the intent is really to requeue. This option should
assist in preventing unintentional cancellations by requiring the user
to explicitly request cancelling scrontab jobs.

Bug 14212