Avoid calling pthread_cancel() on an already-joined thread.
The previous implementation of _wait_for_io() created a _kill_thr() then
used pthread_join() to wait for the I/O thread to finish. The _kill_thr()
sleeps for 300s then issues pthread_cancel() against the I/O thread id.

However, this introduced a race - if the pthread_join() succeeds before
the pthread_cancel() is triggered in the _kill_thr(), and the rest of the
slurmstepd termination hasn't completed before the _kill_thr() sleep is
finished, then we pthread_cancel() an invalid thread id. Which is undefined
behavior that will segfault on some systems.

Instead of that, make _io_thr() detached and use slurmd_cond_timedwait()
with a 300s timeout.

Bug 16063