Remove SLURM_WORKING_CLUSTER env from common job env
If a job lands on a compute node it will be talking to the right cluster
from the slurm.conf on the node. salloc is the only one that needs to
set it so that srun knows which cluster to talk to if it's different
from the local cluster.

Bug 16226