Remove slurm_strftime() macro.
Inline it in the single location it was used, which will also save a
strlcpy() pass on the returned string.

The comment describing this was subtly wrong before - if the size provided
was < 256, and the strftime() result was < 256, then the buffer would be
truncated by this macro as part of the strlcpy(). The refactored code here
instead will always output "####..." instead if the destination buffer
is too small.

This should be nearly impossible in practice - only format strings that
have been pathogically constructed, such as "%c%c%c%c%c", should cause
this anyways.