Fix cgroup cpusets and task affinity if configured < hardware threads
If configured ThreadsPerCore is less than the hardware ThreadsPerCore
hardware, the cgroup cpusets and task affinity was not set correctly.
For example, if configured ThreadsPerCore=1 but the hardware
ThreadsPerCore=2, then Slurm would set cgroup cpusets and bind tasks to
half the cores but both threads on each core.

Fix the outputs hw_sockets, hw_cores, and hw_threads of _get_avail_map()
to be the actual hardware values, not the configured values. Fix the
available CPU bitmap to contain the correct offsets of the configured
CPUs. Fix the task/cgroup cpuset to also contain the correct offsets of
the configured CPUs (which is obtained from xcpuinfo_abs_to_mac()).

Bug 16351