switch/hpe_slingshot - Remove unused vni_pids option.
ALT-225: remove SLINGSHOT_INTER_VNI_PIDS support

Remove the support for the deprecated Slingshot VNI PIDs
option, which was deprecated in an earlier mod.
Keep the 'vni_pids' field in slingshot_jobinfo_t, and pack/unpack
it (should always be NULL) so we don't break the protocol.

Tested on sawmill.

Bug 16768