Fix parentheses at the beginning of square brackets in feature request
If parentheses were at the beginning of square brackets in a feature
request, then they were treated as AND or OR depending on what came before
the square brackets. This is incorrect as anything inside of square brackets
is either XAND or MOR. This resulted in doing a bit_or() or bit_and() of the
feature expression inside the parentheses which could result in incorrect
available nodes.

Bug 16259