Fix srun < 23.02 always getting an "exact" core allocation.
This is due to ntasks_per_core not existing in the step request for
clients older than version 23.02, so ntasks_per_core == 0. This means
that in _pick_step_cores() ntasks_per_core != INFINITE16 is true
(because it is zero), so this calculation results in zero:

tasks_per_node = cores * ntasks_per_core * sockets;

With tasks_per_node == 0 we will never enter the if block that has
use_all_cores = true, so we get use_all_cores = false.

This means that all srun calls are bound to exactly the number of cores
that they request and never the whole job allocation. However, the
number of CPUs shown as allocated in slurmctld/slurmdbd is whatever the
request was (whole or exact), meaning a "whole" step request would only be
bound to "exactly" the number of cores requested but would block all other
steps from using the additional cores.

This is especially problematic for mpirun, which launches a single orted
task per node like this:

srun --ntasks-per-node=1 --ntasks=<num_nodes> orted

Then orted launches all of the requested tasks. This means that the srun
requests only one core per node. With use_all_cores == false, Slurm gives
only a single core for orted. With task/cgroup and ConstrainCores=yes,
Slurm restricts the orted task to only one core.

This is very similar to the problem fixed in bug 13351.

Bug 16333