sacctmgr - Refactor warning thread to avoid pthread_cancel().
There's a subtle race we need to worry about here, especially since
notice_thread_init() and notice_thread_fini() are potentially called
repeatedly within the same process. That race is on notice_thread_init()
followed closely by notice_thread_fini(), and before _print_lock_warn()
has a chance to acquire the mutex. If not for the warn_needed handling,
we'd potentially end up with a stray thread that prints the warning later.