Make _thread_per_group_rpc() threads attached.
The way we use pthread_kill() here means these shouldn't be treated
as detached. If the thread managed to exit before pthread_kill() was
sent it could lead to a segfault - forcing it to stay attached and
ensuring the pthread_join() happens after any other pthread calls
are made should avoid that issue.