When a step requests cpus_per_tres, do not use homogoneous cpus_per_task
If cpus_per_tres is requested, the number of cpus per task on each node
may differ, so we cannot use a homogoneous value for cpus_per_task.
_calc_cpus_per_task() is expected to return 0 for situations where
tasks are not evenly distributed on nodes. This prevents the following
logic _pick_step_nodes() that would be incorrect for cpus_per_tres where
cpus_per_task is not equal on all nodes:

if (cpus_per_task > 0) {
	avail_tasks /= cpus_per_task;
	total_tasks /= cpus_per_task;

Bug 15632