Allow root assoc id != 1 in as_mysql_reset_lft_rgt().
This previously assumed the root assoc id is 1. This is normally true,
however it is possible for the assoic id to be greater than 1
(see bugs 13358 and 15964). This patch lets the reset use the existing
root acct assoc id (even if it is not 1) when recalculating lft/rgt
values.

Bug 16338