Demote error() to SelectType DebugFlag in dist_tasks_gres_min_cpus()
The situation where gres_min_cpus is lower than avail_cpus can happen in the
current code state is quite often. The value of gres_min_cpus is originally
calculated in gres_select_filter_sock_core when we may not know how many
tasks will run on the node. The number or required sockets is caluclated
there to run as many tasks as possible on the node.

The overall logic may be improved which may result in the condition being
unexpected, however, this won't happen on tagged branch and we don't want to
show error while it may be correct behavior.

Bug 16297