Fix communication with srun running outside of a Slurm node.
Regression in 11712ce709.

If we fail to find the node in the configuration we should fallback to
using the raw host value we have in step_ptr.

This may happen if the configured NodeHostname differs from the
gethostname() result on the host, but matches one of the aliases.

Another common case is for srun running on a submit host that is not
configured as a Slurm node.

Bug 16401