Fix shared gres for --tres-per-task with multiple sockets
In _pick_specific_topo(), gres_per_node is used as a total GRES requirement
on the whole node. Previously it was incorrectly calculated as per-socket.

Now, if using --tres-per-task, gres_per_node should be correctly calculated
based on the provided --tres-per-task value and the current node's
pre-determined task count.

Bug 16483