srun - don't set *_PACK_GROUP0 variables on non-HetSteps
The commit applies the same logic to HetSteps as we use for HetJobs.
If we really have a heterogeneous step we set environemnt variables with
_GROUP_PACKX suffix where X is the HetStep offset. However, if we have
only one step (homogeneous step) we use the variable without the suffix.
This way we'll appropriately override variables like SLURM_MEM_PER_NODE
if we call salloc->srun->salloc multiple times with different --mem
requested.

Bug 16240