Remove error path for slurm_cred_send_gids_enabled().
The slurmd will have already filled in the gids if necessary. This
check could inadvertently cause a launch failure if send_gids was
disabled on the slurmctld but left enabled in the slurmd config.

There's no point in crashing the step launch here - we can proceed
safely with the slurmd info if the slurmctld didn't populate it.