Remove exkey handling.
This provided a limited mechanism for key rotation, and was originally
meant for the (long-removed) cred/openssl plugin.

For cred/munge, this could theoretically permit rotation between MUNGE
sockets on the host (through the AuthInfo=socket=/path/to/socket config
option). But there's no support for reconfiguring the auth plugin in
slurmd, so this cut-over would not work reliably in practice.