Fix SLURM_TASKS_PER_NODE
When setting up environment varaibles for salloc/sbatch,
SLURM_TASKS_PER_NODE is set based on the number of tasks in the job.
If the number of tasks is not explicitly given, then we calculate it:
num_tasks = allocated cpus_per_node / requested cpus_per_task.

This calculation is not always correct. For example, --ntasks-per-core=1
or --exclusive can make this calculation wrong.

--ntasks-per-node used to set ntasks implicitly in 22.05, but this was
changed by commits ef51302 and 57ec3eb in 23.02, which meant that
SLURM_TASKS_PER_NODE in 23.02 was now relying on a potentially wrong value
of num_tasks.

So, for --ntasks-per-node we need to calculate
num_tasks = --ntasks-per-node * allocated node count.

--ntasks-per-tres is not an issue because it implicitly sets ntasks
like --ntasks-per-node used to do.
--ntasks-per-socket is also not an issue because it requires --ntasks in
order to work.

SLURM_TASKS_PER_NODE is used by mpirun so it must be set correctly.

Bug 16426

Signed-off-by: Tim McMullan <mcmullan@schedmd.com>