Ensure auth and hash plugins are loaded.
These will usually be triggered by slurm_conf_init_load(), but
that function returns early for a SPANK prolog/epilog that is
using slurm_init() to get access to additional API calls.

Longer-term that locations should be removed in favor of this
only, but for 23.02 this is sufficient to fix it and least
invasive. Expect further refactoring on master.

Bug 16432.

Signed-off-by: Danny Auble <da@schedmd.com>