Bypass plugin setup in slurm_acct_storage_init() when no plugin defined.
Needed to allow slurm_init() to succeed in the SPANK prolog/epilog, as
the pack_slurm_conf_lite() will have left accounting_storage_type
set to NULL. Plugins should not be attempting to contact the database
here, so leave the plugin uninitialized.

This mirrors refactoring for PLUGIN_NOOP already done on 23.02, and this
commit should be thrown away on the merge into master.

Bug 16432.

Signed-off-by: Danny Auble <da@schedmd.com>