Fix nodes having resources allocated from a step after the job completes
If _step_alloc_lps() failed to allocate resources for a step, then resources
allocated to the step were left on the node even after the job completed.
This was a regression caused by commit ff5b5623c4.

This commit fixes the problem solved by commit ff5b5623c4 in a different
way. Rather than restore job_resources from a backup, allow
_step_alloc_lps() to allocate resources for the step on all nodes even if
one of the nodes fails. Then, deallocate all resources that were allocated
to that step if there was a failure.

Bug 14263