Fix salloc/sbatch/srun setting ntasks incorrectly
salloc, sbatch, and srun could set ntasks incorrectly with
-B,--extra-node-info or one of the specific options:
--sockets-per-node
--cores-per-socket
--threads-per-core (or --hint=nomultithread)

When these options are specified and ntasks is not, then --ntasks is set
implicitly. However, this did not account for --ntasks-per-node.
This commit avoids implicitly setting --ntasks if --ntasks-per-node is
given. This is in line with the behavior from commits ef51302 and
57ec3eb.

Bug 16666