slurmctld - handle negated nodelist after values have been updated
Move the latest ExcNodeList update, the one based on negated NodeList
merge, to an independent place far from the original point; where it is
ensured that job's ExcNodeList, NodeList, and MinNodes have their final
values.

We want to ensure that the negated NodeList merge is done based on the
job's final configuration, no matter the update flags that have been
sent.

Bug 16573