slurmctld - check min_nodes against the final value set for the job
So in _update_job, after the previous commit, we no longer need to check it
against the maybe sent, maybe not, or maybe not even the final value,
that the update job RPC struct will provide.

We can now check it against the min_nodes from the job records. At this
point in code, for this function, we already have the final value stored.

Bug 16573