switch/hpe_slingshot - Aovid error message when freeing the same vni.
When het steps or het jobs complete, the switch plugin is called
separately for each component. This can lead to freeing the same VNI
multiple times, and warning messages in the log. There's not enough
information available in the function parameters to determine if it's
part of a het step or job, so the best we can do (without changing
APIs or data structures) is to change the warning messages to debug.

Bug 16590