burst_buffer/lua - always queue_job_scheduler at the end of stage-in
Previously this was only done if "capacity" was requested as part of the
burst buffer directive.

Without calling queue_job_scheduler(), the job will still be scheduled by
the backfill scheduler, or if another event happens that queues the main
scheduler (such as job submission or completion). But, this commit allows
the job to be scheduled sooner.

Bug 16715