Fix slurmrestd handling of job hold/release operations.
Slurmrestd uses the same parser handler for the 'hold' parameter for job
submit/update. This requires a dedicated handling on slurmctld side, since
the case '"hold": false' is translated as job_desc.priority=INFINITE for
job updates to denote the release request of a held job, and thus for
job submission INFINITE wasn't originally handled as it wasn't needed
for CLI tools, but now it is for slurmrestd.

The commit adds the option to data_parser - v39 was totally missing the
support before and v38 support was buggy in terms of submission of a job
from SlurmUser with '"hold": false'.

Bug 16582