Fix TresUsageIn[Tot|Ave] calculation for gres/gpumem and gres/gpuutil.
TresUsageInTot for gpumem/gpuutil represents the sum of the high watermarks
of all the ranks within the step. By default, tres_usage_in_tot contains the
last recorded usage, which is not what we want to sum up among the ranks.

This commit changes how the total is calculated for gpumem/gpuutil by
overwriting the last recorded usage in tres_usage_in_tot with the high
watermark that is stored in tres_usage_in_max, similar to what is already
done for mem/vmem.

Bug 16619