Skip deallocating resources for the batch, extern, and interactive steps
Resources allocated to these steps are not added to the used job
resources. So, there is no need to deallocate resources used by these
steps.

Remove code that was trying to skip deallocating resources by checking
if (step_layout == NULL). step_layout is not NULL for these
steps because of fake_slurm_step_layout_create().

Bug 15632