Allow slurmctld to kill steps when RESPONSE_JOB_STEP_CREATE msg fails
Previously the controller was not checking wether the job_step_create
RPC successfully replied back to srun with RESPONSE_JOB_STEP_CREATE msg.
This could lead into a state in which the controller allocated
resources for the step, but srun was not aware of them, resulting in a
failed/stuck step, but with ghost resources allocated indefinitely. This
was preventing further steps in the same job from starting (due to busy
resources).

New behavior checks if RESPONSE_JOB_STEP_CREATE was successfully sent,
and kills the step on the controller side if the message failed.

Bug 16714