Fix parentheses and matching OR logic with changeable features
Continuation of commit b5322c6900c, but that commit only fixed feature
expressions with static features.

When changeable features are requested and a bar operator is used, the bar
operator is always interpreted as matching OR (not boolean OR), and square
brackets are not explicitly requested. By implying bracket==1 on all
features except for the last one, this commit uses the fix from commit
b5322c6900c.

Even though a bracket is implied, '&' is still always treated as normal
AND, not XAND.

Bug 17111