srun - Continue waiting for allocation on SRUN_JOB_COMPLETE job_id mismatch
Or on any unexpected message types.

While srun is waiting for the allocation, avoid terminating the step launch if
the message was not intended for us.

Receiving SRUN_JOB_COMPLETE for a different job_id is possible (or other
unexpected message typesas well) if a port in SrunPortRange was reused
recently and the slurmctld is delayed communicating with the srun
that previously occupied that port.

Bug 16770