Fix reservation accessibility for associations composed of partition(s)
When specifying a partition in the reservation creation/update we are not
interested in restricting its access, but rather to specify that nodes
reserved by the reservation should be the ones in the partition.

The bug was preventing to create/update reservations specifying user and
account when the existing association in the database also had partition(s).

The commit makes it so that at reservation creation/update time the user and
account are checked to be a matching association but ignoring the
partition part of the association.

Bug 16497