Do not use threads per core when deallocating steps with cr_cpu.
This fixes a cpu underflow error with cr_cpu and --threads-per-core.

With SelectTypeParameters=cr_cpu and a step requesting --threads-per-core
< cpus_per_core, this error would happen when the step was deallocated:

 error: _step_dealloc_lps: CPU underflow for JobId=1468 StepId=0 \
						   (16<32 on job node 0)

This happened because _step_alloc_lps() uses threads per core when
setting the allocated cpu count only with cr_core and cr_socket, but
_step_dealloc_lps() used threads per core regardless of cr_type.
This commit ensures that _step_alloc_lps() and _step_dealloc_lps()
do the same thing.

Bug 16811