Make job_desc->min_cpus accurate before job_submit is run.
min_cpu calculations are largely based on ntasks, and commit 57ec3eb55d
removed the calculation of ntasks based on ntasks-per-node in the client
commands. This means job_desc->min_cpus wasn't an accurate representation
of the job request by the time job submit was run in some cases.

This commit makes job_desc->min_cpus accurate at the time of the job_desc
creation, so it will be ready in time for the job_submit to use in
valid_job_create_req().

Bug 16744