Fix regression that doesn't terminate jobacct_gather plugin
After bug 01eb47fe2b the acct_gather_profile was made a noop, but still
would start the jobacct_gather plugin for polling stats and to make test
12.2 happy.

Since the plugin was a noop, it didn't start any g_context, and when fini
was called it effectively did nothing, including not terminating the
jobacct_gather plugin. This left the system with an uncleaned cgroup tree.

Force the jobacct_gather termination as we force the jobacct_gather poll
mechanism to take place.

Bug 16916