Fix job update to comma-separated list of reservations
Updating job reservations job_ptr->resv_list was updated correctly, but
job_ptr->resv_name was set to only one of the names. This resulted in
job that was able to run in any of the reservations used in the update
request, while `scontrol show job` shows only one reservation. And if
the controller was restarted, the list of reservations would be lost and
only the one reservation in resv_name would be restored.

Bug 16939