Remove JOB_RESIZED flag in favor of local variable.
The JOB_RESIZED flag is no longer needed after commit 782d1c1939.
Originally it indicated that the step's core bitmap and the job's core
bitmap differed after removing a node from a job with an active job step.
But this is now handled immediately, so the flag is not necessary.

Bug 16959