Docs - update Prolog and Epilog, and enable_configless descriptions.
Configless now can have the slurmctld push Prolog and Epilog scripts to
slurmds with the limitation that the value of Prolog and Epilog are
filenames that are located in the same directory as slurm.conf. As a result
Prolog and Epilog no longer need to be fully qualified paths.

Bug 16876