Fix parsing of max_submit_line_size SchedulerParameter
The integer value of max_submit_line_size SchedulerParameter was not
properly parsed, it was always set to 0 when defined in configuration
file. This was due to wrong string offset when converting the integer
value. The string max_submit_line_size= is 21 bytes long, this commit
changes this offset everywhere this scheduler parameter is parsed.

Bug 17015.