Allow configless mode to push Prolog and Epilog scripts to slurmds.
Allow for Prolog and Epilog scripts to be pushed to slurmds when
enable_configless is specified, and the Prolog and Epilog are non-absolute
path.

This change also allows for the pathnames of the Prolog and Epilog scripts
to be relative to the slurm.conf directory when enable_configless is not
specified.

Bug 16876