Docs - Fix manuals for preempt/qos
The case of PreemptType=preempt/qos and PreemptMode=GANG,SUSPEND is
structurally broken in the current code state. It may lead to incorrect
PreemptMode being used or wrong job (lower QoS can preempt higher QoS).

The reason is that _job_test makes use of GANG logic to suspend the
preemptor. However it selects resources based on partition bitmaps (it's
not aware of QoS) and relies on GANG scheduler to RESUME higher
PriorityTier job and case shadow on finally preempted job.

Bug 16441