scontrol - fix comma parsing in scontrol_process_plus_minus()
Fix scontrol_process_plus_minus() to correctly handle commas in node strings
that are assumed to be contained within brackets. For example
"h[1,5],h[8,9]" will be split on the comma before the h only so that the
remaining commas may be processed by hostlist functions. Previously, splits
happened on all commas regardless of context.

Bug 16758