slurmctld - fix comma parsing when processing reservation node change
When _get_update_node_bitmap() processed a node string, it called strtok_r()
with a comma delimiter to tokenize the string. Given a node string
like this:
        h[1,2],h[3,4]

The string was improperly split into the tokens "h[1", "2]", "h[3", "4]". To
resolve the problem nodestring_comma_strtok_r() is called instead which will
tokenize only on commas not followed by a digit. So, the example string is
correctly split into the tokens "h[1,2]" and "h[3,4]".

Bug 16758