Fix perlapi build when using non-default libdir
Previously, Makefile.PL.in used @prefix@/lib to find the Slurm libraries
when compiling. The problem is that if --libdir is used at configure
time then the Slurm libraries wouldn't be found.

Configure creates a Makefile.PL from the Makefile.PL.in and replaces
@libdir@ with the "libdir" variable from configure. If --libdir is
specified with configure then libdir is the given path. If --libdir is
not specified, then libdir is defined as "${exec_prefix}/lib".

The libslurm[db]/Makefile runs "perl Makefile.PL" and generates a
Makefile in the perl directory using ExtUtils::MakeMaker. This Makefile
compiles against the Slurm libraries using the specified "rpath".

If --libdir isn't used at configure time and "${exec_prefix}/lib" is
used in Makefile.PL and it is enclosed in double quotes, then perl will
try interpolating ${exec_prefix} as a variable and will result in a
blank string leaving "/lib" in the resulting Makefile's rpath and not
being able to find the Slurm libraries.

The solution is to change the double quotes to single quotes in
Makefile.PL so that perl doesn't interpolate the variable and leaves
${exec_prefix} in place in the resulting Makefile and to pass
exec_prefix to the Makefile to satisfy $exec_prefix.

Bug 9109

Co-authored-by: Felix-Antoine Fortin <felix-antoine.fortin@calculquebec.ca>